feat: export to umd file
