+++ package.json
@@ -36,6 +36,7 @@
     "@graphiql/toolkit": "^0.8.0",
     "@rollup/plugin-commonjs": "^23.0.0",
     "@rollup/plugin-node-resolve": "^15.0.0",
+    "@svgr/rollup": "6.4.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
@@ -80,6 +81,7 @@
   },
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
+  "umd": "dist/umd/index.js",
   "files": [
     "dist"
   ]
